pyproject.toml
[tool.poetry]
name = "Research Assistance"
version = "0.1.0"
description = ""
authors = ["devcode8 <devccode8@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fetchai = "^0.1.30"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
langchain = "^0.3.15"
langchain-openai = "^0.3.1"
langchain-community = "^0.3.15"
wikipedia = "^1.4.0"
arxiv = "^2.1.3"
openai = "^1.60.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
